validation/requirements.txt
pytest
nose2
PyYaml